requirements.txt
pandas
requests
apache-airflow-providers-slack
